# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Factur-IA API",
    "version": "1.5.0"
  },
  "paths": {
...
    "/abonnements/": {
      "get": {
        "tags": [
          "Gestion des Abonnements"
        ],
        "summary": "List Plans",
        "description": "Liste tous les plans d'abonnement disponibles sur la plateforme.\nAccessible par tout utilisateur authentifié.",
        "operationId": "list_plans_abonnements__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AbonnementRead"
                  },
                  "type": "array",
                  "title": "Response List Plans Abonnements  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Gestion des Abonnements"
        ],
        "summary": "Create Plan",
        "description": "Crée un nouveau plan d'abonnement.\nRéservé aux administrateurs de la plateforme.",
        "operationId": "create_plan_abonnements__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbonnementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbonnementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/abonnements/me": {
      "get": {
        "tags": [
          "Gestion des Abonnements"
        ],
        "summary": "Get My Subscriptions",
        "description": "Récupère les abonnements des entreprises (espaces de travail)\nauxquelles l'utilisateur actuel est rattaché.",
        "operationId": "get_my_subscriptions_abonnements_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntrepriseAbonnementRead"
                  },
                  "type": "array",
                  "title": "Response Get My Subscriptions Abonnements Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/abonnements/{abonnement_id}": {
      "patch": {
        "tags": [
          "Gestion des Abonnements"
        ],
        "summary": "Update Plan",
        "description": "Modifie les caractéristiques d'un plan d'abonnement.\nRéservé aux administrateurs de la plateforme.",
        "operationId": "update_plan_abonnements__abonnement_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "abonnement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Abonnement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbonnementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbonnementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Gestion des Abonnements"
        ],
        "summary": "Delete Plan",
        "description": "Supprime un plan d'abonnement.\nRéservé aux administrateurs de la plateforme.",
        "operationId": "delete_plan_abonnements__abonnement_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "abonnement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Abonnement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/utilisateurs/": {
      "get": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "List Team Members",
        "description": "Liste tous les utilisateurs rattachés à l'entreprise active (l'équipe).",
        "operationId": "list_team_members_utilisateurs__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UtilisateurRead"
                  },
                  "title": "Response List Team Members Utilisateurs  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Create Team Member",
        "description": "Crée un nouvel utilisateur et le rattache automatiquement à l'entreprise actuelle.",
        "operationId": "create_team_member_utilisateurs__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/utilisateurs/{user_id}": {
      "delete": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Delete Team Member",
        "operationId": "delete_team_member_utilisateurs__user_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Update Team Member",
        "description": "Modifie les informations d'un\ncollaborateur (incluant ses rôles et accès admin).",
        "operationId": "update_team_member_utilisateurs__user_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurTeamUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admins-plateforme/": {
      "get": {
        "tags": [
          "Administration Plateforme"
        ],
        "summary": "List Admins",
        "description": "Liste les administrateurs de la plateforme.",
        "operationId": "list_admins_admins_plateforme__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminPlateformeRead"
                  },
                  "type": "array",
                  "title": "Response List Admins Admins Plateforme  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admins-plateforme/{utilisateur_id}/promouvoir": {
      "post": {
        "tags": [
          "Administration Plateforme"
        ],
        "summary": "Promote Admin",
        "description": "Promeut un utilisateur au rang d'administrateur de plateforme.\nNe modifie que le statut plateforme, jamais les rôles métier.",
        "operationId": "promote_admin_admins_plateforme__utilisateur_id__promouvoir_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "utilisateur_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Utilisateur Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPlateformeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admins-plateforme/{utilisateur_id}/revoquer": {
      "post": {
        "tags": [
          "Administration Plateforme"
        ],
        "summary": "Revoke Admin",
        "description": "Révoque le statut d'administrateur de plateforme d'un utilisateur.\nLe compte reste inchangé par ailleurs (rôles, entreprises, existence).",
        "operationId": "revoke_admin_admins_plateforme__utilisateur_id__revoquer_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "utilisateur_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Utilisateur Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPlateformeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbonnementCreate": {
        "properties": {
          "libelle": {
            "type": "string",
            "maxLength": 100,
            "title": "Libelle",
            "description": "Nom du plan (ex: Pro, Business)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tarif": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Tarif",
            "default": "0"
          },
          "nombre_max_utilisateurs": {
            "type": "integer",
            "title": "Nombre Max Utilisateurs",
            "default": 1
          },
          "nombre_max_factures_mois": {
            "type": "integer",
            "title": "Nombre Max Factures Mois",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "libelle"
        ],
        "title": "AbonnementCreate",
        "description": "Schéma pour la création d'un nouveau plan (Admin uniquement)."
      },
      "AbonnementRead": {
        "properties": {
          "libelle": {
            "type": "string",
            "maxLength": 100,
            "title": "Libelle",
            "description": "Nom du plan (ex: Pro, Business)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tarif": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)",
            "title": "Tarif",
            "default": "0"
          },
          "nombre_max_utilisateurs": {
            "type": "integer",
            "title": "Nombre Max Utilisateurs",
            "default": 1
          },
          "nombre_max_factures_mois": {
            "type": "integer",
            "title": "Nombre Max Factures Mois",
            "default": 10
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "libelle",
          "id"
        ],
        "title": "AbonnementRead",
        "description": "Schéma pour la lecture d'un plan d'abonnement."
      },
      "AbonnementUpdate": {
        "properties": {
          "libelle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Libelle"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tarif": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tarif"
          },
          "nombre_max_utilisateurs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Max Utilisateurs"
          },
          "nombre_max_factures_mois": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Max Factures Mois"
          }
        },
        "type": "object",
        "title": "AbonnementUpdate",
        "description": "Schéma pour la modification d'un plan (Admin uniquement)."
      },
...
      "EntrepriseAbonnementRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_entreprise": {
            "type": "integer",
            "title": "Id Entreprise"
          },
          "id_abonnement": {
            "type": "integer",
            "title": "Id Abonnement"
          },
          "date_debut": {
            "type": "string",
            "format": "date",
            "title": "Date Debut"
          },
          "date_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Fin"
          },
          "statut": {
            "$ref": "#/components/schemas/StatutSouscription"
          },
          "date_creation": {
            "type": "string",
            "format": "date-time",
            "title": "Date Creation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_entreprise",
          "id_abonnement",
          "date_debut",
          "date_fin",
          "statut",
          "date_creation"
        ],
        "title": "EntrepriseAbonnementRead",
        "description": "Détails de la souscription d'une entreprise à un abonnement."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UtilisateurCreate": {
        "properties": {
          "nom": {
            "type": "string",
            "maxLength": 255,
            "title": "Nom"
          },
          "prenom": {
            "type": "string",
            "maxLength": 255,
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "type": "boolean",
            "title": "Est Actif",
            "default": true
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Mot de passe en clair"
          },
          "id_role": {
            "type": "integer",
            "title": "Id Role",
            "description": "ID du rôle métier rattaché"
          },
          "est_admin": {
            "type": "boolean",
            "title": "Est Admin",
            "description": "Droit admin au niveau de l'entreprise",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "nom",
          "prenom",
          "email",
          "password",
          "id_role"
        ],
        "title": "UtilisateurCreate",
        "description": "Schéma pour créer un utilisateur avec son mot de passe en clair."
      },
      "UtilisateurRead": {
        "properties": {
          "nom": {
            "type": "string",
            "maxLength": 255,
            "title": "Nom"
          },
          "prenom": {
            "type": "string",
            "maxLength": 255,
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "type": "boolean",
            "title": "Est Actif",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date_creation": {
            "type": "string",
            "format": "date-time",
            "title": "Date Creation"
          },
          "date_modification": {
            "type": "string",
            "format": "date-time",
            "title": "Date Modification"
          },
          "date_derniere_connexion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Derniere Connexion"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "est_admin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Admin",
            "description": "Statut administrateur du membre dans l'entreprise active. Lecture seule : à utiliser pour pré-remplir le formulaire d'édition et éviter de retirer les droits par erreur."
          }
        },
        "type": "object",
        "required": [
          "nom",
          "prenom",
          "email",
          "id",
          "date_creation",
          "date_modification"
        ],
        "title": "UtilisateurRead",
        "description": "Schéma de sortie (le hash n'est jamais renvoyé)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OCR-Secret-Token"
      }
    }
  }
}